Fixed rating appearing as a double in myReviews, now properly appears as an int.

+++ src/main/java/edu/virginia/sde/reviews/MyReviewObject.java
@@ -15,7 +15,7 @@ public class MyReviewObject{
        this.rating = rating;
     }
 
-    public double getRating() {
+    public int getRating() {
         return rating;
     }
